+++ workspace/recursos/src/main/resources/sql/create_nota_credito.sql
@@ -76,7 +76,7 @@ create table NOTA_CREDITO_DETALLE
     REFERENCIA_ID BIGINT,
     TIPO_REFERENCIA varchar(5),
     NOTA_CREDITO_ID BIGINT,
-    CANTIDAD integer ,
+    CANTIDAD decimal(13,5),
     PRECIO_UNITARIO decimal(13,3),
     DESCUENTO decimal(13,2),
     VALOR_ICE decimal,
